feat: 필터링 닫기 구현
