+++ web/src/components/BarrierFreeMap/CategoryList/CategoryList.styles.ts
@@ -8,6 +8,7 @@ export const layout = css({
 
 export const buttonList = css({
   display: 'flex',
+  alignItems: 'center',
   gap: '0.5rem',
 
   maxW: 'calc(100vw - 4rem)',
@@ -19,35 +20,23 @@ export const buttonList = css({
   },
 });
 
-export const defaultButton = css({
-  p: '0.4rem 1.4rem',
-  border: '1px solid {colors.primary.500}',
+export const cancelButton = css({
+  p: '0.4rem',
+  border: '1px solid {colors.grey.400}',
   rounded: 'full',
 
   bg: '{colors.white}',
-  color: '{colors.primary.800}',
+  color: '{colors.grey.400}',
   textStyle: 'body',
 
   cursor: 'pointer',
   transition: 'all 0.2s',
 
   _hover: {
-    bg: '{colors.primary.100}',
+    bg: '{colors.grey.100}',
   },
 
   _active: {
-    bg: '{colors.primary.100}',
+    bg: '{colors.grey.100}',
   },
-});
-
-export const activeButton = css({
-  p: '0.4rem 1.4rem',
-  border: '1px solid {colors.primary.500}',
-  rounded: 'full',
-
-  bg: '{colors.primary.500}',
-  color: '{colors.white}',
-  textStyle: 'body',
-
-  cursor: 'pointer',
 });
