+++ web/src/components/BarrierFreeMap/CategoryList/CategoryList.tsx
@@ -1,17 +1,13 @@
 import { useState } from 'react';
+import { IoCloseOutline } from 'react-icons/io5';
 
 import type { MarkerType } from '@/pages/BarrierFreeMap/BarrierFreeMap.type';
 
-import * as styles from './CategoryList.styles';
+import CategoryButton from '@/components/BarrierFreeMap/CategoryButton/CategoryButton';
 
-const MARKER_TYPE: Record<MarkerType, string> = {
-  elevator: '엘리베이터',
-  impariment_toilet: '장애인 화장실',
-  ramp: '경사로',
-  note: '특이 사항 존재',
-  audio_device: '음성 유도기',
-  support_office: '장애학생지원실',
-};
+import { MARKER_TYPE } from '@/constants/barrierFree';
+
+import * as styles from './CategoryList.styles';
 
 interface CategoryListProps {
   categories: Record<MarkerType, boolean>;
@@ -21,25 +17,30 @@ interface CategoryListProps {
 const CategoryList = ({ categories, filterCategories }: CategoryListProps) => {
   const [isOpen, setIsOpen] = useState(false);
 
+  const openList = () => setIsOpen(true);
+  const closeList = () => setIsOpen(false);
+
   return (
     <div className={styles.layout}>
       {isOpen ? (
         <ul className={styles.buttonList}>
+          <li>
+            <button className={styles.cancelButton} onClick={closeList}>
+              <IoCloseOutline size="2rem" />
+            </button>
+          </li>
           {Object.entries(MARKER_TYPE).map(([key, value]) => (
             <li key={key}>
-              <button
-                className={categories[key as MarkerType] ? styles.activeButton : styles.defaultButton}
+              <CategoryButton
+                value={value}
+                isActive={categories[key as MarkerType]}
                 onClick={() => filterCategories(key as MarkerType)}
-              >
-                {value}
-              </button>
+              />
             </li>
           ))}
         </ul>
       ) : (
-        <button className={styles.defaultButton} onClick={() => setIsOpen(true)}>
-          필터 적용하기
-        </button>
+        <CategoryButton value="필터 적용하기" onClick={openList} />
       )}
     </div>
   );
